Validation Checks
Made it so that when a Manager wishes to add/edit an employee they must make sure the password is secure (6-20 characters, 1 number, 1 special character) and that the username/email they have input is NOT someone else's. This check is also made during registration as well as on the manager's personal page.